Remove old comments
